refactor(adminuser): 将ID字段类型由int改为int64
- 将AdminUser实体中的ID字段类型修改为int64以支持更大范围的ID值
- 更新了所有相关方法和函数的ID参数类型，包括查询、删除、更新和创建操作
- 调整了生成的SQL规格中的ID字段类型为int64
- 修改了AdminUserMutation中的ID相关字段和方法的类型
- 更新了验证器的签名以匹配int64类型
- 迁移方案中id列的数据类型改为int64，确保数据库字段一致性

+++ ent/adminuser.go
@@ -17,7 +17,7 @@ type AdminUser struct {
 	config `json:"-"`
 	// ID of the ent.
 	// 自增主键ID
-	ID int `json:"id,omitempty"`
+	ID int64 `json:"id,omitempty"`
 	// 用户名
 	Username string `json:"username,omitempty"`
 	// 密码哈希值
@@ -74,7 +74,7 @@ func (_m *AdminUser) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			_m.ID = int(value.Int64)
+			_m.ID = int64(value.Int64)
 		case adminuser.FieldUsername:
 			if value, ok := values[i].(*sql.NullString); !ok {
 				return fmt.Errorf("unexpected type %T for field username", values[i])
